+++ packages/agent/src/cbor.ts
@@ -4,7 +4,7 @@ import borc from 'borc';
 import * as cbor from 'simple-cbor';
 import { CborEncoder, SelfDescribeCborSerializer } from 'simple-cbor';
 import { Principal } from '@dfinity/principal';
-import { concat, fromHex } from './utils/buffer';
+import { concat, fromHex, toHex } from './utils/buffer';
 
 // We are using hansl/simple-cbor for CBOR serialization, to avoid issues with
 // encoding the uint64 values that the HTTP handler of the client expects for
@@ -125,5 +125,9 @@ export function decode<T>(input: ArrayBuffer): T {
     },
   });
 
-  return decoder.decodeFirst(buffer);
+  try {
+    return decoder.decodeFirst(buffer);
+  } catch(e: unknown) {
+    throw new Error(`Failed to decode CBOR: ${e}, input: ${toHex(buffer)}`);
+  }
 }