+++ docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- feat: make `IdbStorage` `get/set` methods generic
+
 ## [1.2.0] - 2024-03-25
 
 ### Added