+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - feat: make `IdbStorage` `get/set` methods generic
 - chore: add context to errors thrown when failing to decode CBOR values.
+- chore: replaces globle npm install with setup-node for size-limit action
 
 ## [1.2.0] - 2024-03-25