+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.2.1] - 2024-04-25
+
 ### Changed
 
 - feat: make `IdbStorage` `get/set` methods generic